requirements.txt
Django==4.1.2
django-environ==0.9.0
django-taggit==3.0.0
coverage==6.5.0
